clear media ans variation metadata cache when an event occurs in the media lifetime

+++ src/Storage/MediaPropertyAccessor.php
@@ -16,7 +16,17 @@ class MediaPropertyAccessor
         private readonly Filesystem $filesystem,
         private readonly MimeTypeGuesser $mimeTypeGuesser,
         private readonly CacheInterface $cache,
+        private readonly int $expiresAfter = 3600 * 24,
     ) {
+    }
+
+    public function clearCache(string $path): void
+    {
+        $this->cache->delete($this->getCacheKey($path, 'filesize'));
+        $this->cache->delete($this->getCacheKey($path, 'format'));
+        $this->cache->delete($this->getCacheKey($path, 'mime_type'));
+        $this->cache->delete($this->getCacheKey($path, 'pixel_dimensions'));
+        $this->cache->delete($this->getLastModifiedCacheKey($path));
     }
 
     public function getMimeType(string $path): string
@@ -29,16 +39,17 @@ class MediaPropertyAccessor
 
     public function getFormat(string $path): string
     {
-        $mimeType = $this->getMimeType($path);
-
-        return $this->mimeTypeGuesser->getPossibleExtension($mimeType);
+        return $this->cache->get(
+            $this->getCacheKey($path, 'format'),
+            fn (ItemInterface $item): string => $this->guessFormat($path),
+        );
     }
 
     public function getFileSize(string $path): int
     {
         return $this->cache->get(
             $this->getCacheKey($path, 'filesize'),
-            fn (ItemInterface $item): int => $this->filesystem->fileSize($path)
+            fn (ItemInterface $item): int => $this->guessFilesize($path)
         );
     }
 
@@ -59,11 +70,15 @@ class MediaPropertyAccessor
     public function getLastModified(string $path): int
     {
         return $this->cache->get(
-            \sprintf('joli_media_property_%s_%s_lastModified', $this->libraryName, Resolver::normalizePath($path)),
+            $this->getLastModifiedCacheKey($path),
             function (ItemInterface $item) use ($path): int {
-                $item->expiresAfter(24 * 60 * 60); // every day, check if the file has been modified
+                $item->expiresAfter($this->expiresAfter);
 
-                return $this->filesystem->lastModified($path);
+                try {
+                    return $this->filesystem->lastModified($path);
+                } catch (UnableToRetrieveMetadata) {
+                    return time();
+                }
             },
         );
     }
@@ -79,11 +94,40 @@ class MediaPropertyAccessor
         );
     }
 
+    private function getLastModifiedCacheKey(string $path): string
+    {
+        return \sprintf(
+            'joli_media_property_%s_%s_lastModified',
+            $this->libraryName,
+            Resolver::normalizePath($path),
+        );
+    }
+
+    private function guessFilesize(string $path): int
+    {
+        try {
+            return $this->filesystem->fileSize($path);
+        } catch (UnableToRetrieveMetadata) {
+            return 0;
+        }
+    }
+
+    private function guessFormat(string $path): string
+    {
+        $mimeType = $this->getMimeType($path);
+
+        return $this->mimeTypeGuesser->getPossibleExtension($mimeType);
+    }
+
     private function guessMimeType(string $path): string
     {
         try {
             $mimeType = $this->filesystem->mimeType($path);
         } catch (UnableToRetrieveMetadata) {
+            if (!$this->filesystem->has($path)) {
+                return 'application/octet-stream';
+            }
+
             // try to guess the mime type from the content
             $mimeType = $this->mimeTypeGuesser->guessMimeTypeFromContent(
                 $this->filesystem->read($path),
@@ -103,6 +147,11 @@ class MediaPropertyAccessor
         }
 
         $temporaryFile = tempnam(sys_get_temp_dir(), 'image');
+
+        if (!$this->filesystem->has($path)) {
+            return false;
+        }
+
         file_put_contents($temporaryFile, $this->filesystem->read($path));
         $imageSize = getimagesize($temporaryFile);
         unlink($temporaryFile);
